Adds examples/graphql.rs
I wrote this example for another project but think it would be useful to contribute to the examples/ folder. It's uses the Juniper library to implement a GraphQL server, and shows an example HTML page that can write GraphQL requests to it.

This probably isn't in the right style (this is written as a cargo-script, for one) but I'm happy to edit it to be a good example for inclusion.