Update rng used to generate session id
Version 5.0 of rand changed the algorithm
behind `ThreadRng` to a cryptographicly secure one.
https://docs.rs/rand/0.5.0/rand/rngs/struct.ThreadRng.html

`ThreadRng` should both be faster and more fault tolerant
than `OsRng`.

Also removed the filter after the generation of random data.
`Alphanumeric` only generates characters in the range 'a'-'z',
'A'-'Z' and '0'-'9', so the filter was a no-op.