Unpin time-rs and bump MSRV to 1.51
As the ecosystem is making no effort to support MSRVs older than about 6
months, keeping a pin on a base crate like time just causes more
problems than it solves.

Bump the version we use in our lints and tests to 1.51 and fix the few
new Clippy items that brings up.