content-encoding: fix Gzip test on non-linux hosts
The output test for Gzip encoding uses a hardcoded expected value,
however the Gzip header is unstable containing both a timestamp and an
OS ID field.

The timestamp field is apparently not set by the `deflate` crate but the
OS ID is, meaning this test consistently fails on systems other than
x86_64 Linux.

Here we just change the test to ignore the 10-byte header, rather than
trying to hack together a proper round-trip test in the absence of a
decoder method in the library crate.